removes files unused
