--- src/app.go
@@ -1,104 +1,0 @@
-package main
-
-import (
-	"errors"
-	"fmt"
-	"log"
-
-	"github.com/awesome-gocui/gocui"
-)
-
-var (
-	viewArr = []string{"usuario", "juego"}
-	active  = 0
-)
-
-func setCurrentViewOnTop(g *gocui.Gui, name string) (*gocui.View, error) {
-	if _, err := g.SetCurrentView(name); err != nil {
-		return nil, err
-	}
-	return g.SetViewOnTop(name)
-}
-
-func nextView(g *gocui.Gui, v *gocui.View) error {
-	nextIndex := (active + 1) % len(viewArr)
-	name := viewArr[nextIndex]
-
-	out, err := g.View("juego")
-	if err != nil {
-		return err
-	}
-	fmt.Fprintln(out, "Paso de vista "+v.Name()+" a "+name)
-
-	if _, err := setCurrentViewOnTop(g, name); err != nil {
-		return err
-	}
-
-	if nextIndex == 0 || nextIndex == 1 {
-		g.Cursor = true
-	} else {
-		g.Cursor = false
-	}
-
-	active = nextIndex
-	return nil
-}
-
-func layout(g *gocui.Gui) error {
-	maxX, maxY := g.Size()
-	if v, err := g.SetView("usuario", 0, 0, maxX/2-1, maxY/2-1, 0); err != nil {
-		if !errors.Is(err, gocui.ErrUnknownView) {
-			return err
-		}
-		v.Title = "Usuario"
-		v.Editable = true
-		v.Wrap = true
-
-		if _, err = setCurrentViewOnTop(g, "usuario"); err != nil {
-			return err
-		}
-	}
-
-	if v, err := g.SetView("juego", maxX/2-1, 0, maxX-1, maxY/2-1, 0); err != nil {
-		if !errors.Is(err, gocui.ErrUnknownView) {
-			return err
-		}
-		v.Title = "juego"
-		v.Wrap = true
-		v.Autoscroll = true
-	}
-
-	return nil
-}
-
-func quit(g *gocui.Gui, v *gocui.View) error {
-	return gocui.ErrQuit
-}
-
-func main() {
-	g, err := gocui.NewGui(gocui.OutputNormal, true)
-	if err != nil {
-		log.Panicln(err)
-	}
-	defer g.Close()
-
-	g.Highlight = true
-	g.Cursor = true
-	g.SelFgColor = gocui.ColorGreen
-
-	g.SetManagerFunc(layout)
-
-	if err := g.SetKeybinding("", gocui.KeyCtrlC, gocui.ModNone, quit); err != nil {
-		log.Panicln(err)
-	}
-	if err := g.SetKeybinding("", gocui.KeyEsc, gocui.ModNone, quit); err != nil {
-		log.Panicln(err)
-	}
-	if err := g.SetKeybinding("", gocui.KeyTab, gocui.ModNone, nextView); err != nil {
-		log.Panicln(err)
-	}
-
-	if err := g.MainLoop(); err != nil && !errors.Is(err, gocui.ErrQuit) {
-		log.Panicln(err)
-	}
-}
